+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 - default to system theme on first launch
+- display URL instead of title on hover for tabs with loading issues
 
 ## [0.0.15] - 2019-10-06